requirements.txt
python-dotenv
langchain-core
langchain-community
langchain
langchain_text_splitters
langchain-chroma
langchain-openai
numexpr
youtube_search
langchain_experimental